Adaptive quad pool operations (#36)
This patch adds modules for adaptive quad pooling operations. It changes
the implementation of the C++ functions: now both indices and values are
saved as values of unordered maps (tile -> index, tile -> value).

Also, adaptive modules now work with sparse tensors, so that backward
pass operation is computed with the same size of weight tensor.